usb: Fix a few more warnings.

+++ pcsx2/USB/usb-eyetoy/cam-windows.cpp
@@ -471,7 +471,7 @@ namespace usb_eyetoy
 		{
 			mpeg_mutex.lock();
 			int len2 = mpeg_buffer.length;
-			if (len < mpeg_buffer.length)
+			if ((unsigned int)len < mpeg_buffer.length)
 				len2 = len;
 			memcpy(buf, mpeg_buffer.start, len2);
 			mpeg_buffer.length = 0;
